requirements.txt
tensorflow==2.12.0
scipy==1.11.3
numpy==1.23.5
dash>=2.14.0
plotly-resampler>=0.9.1
plotly==5.17.0
flask-caching==2.0.2
timm==0.9.8
einops==0.7.0
pytorch==2.1.0